+++ search-films/src/components/Menu/index.tsx
@@ -1,12 +1,23 @@
-import { useContext } from 'react';
-import { IUserContext, UserContext } from '../../context/user.context';
+import { useDispatch, useSelector } from 'react-redux';
+import { RootState } from '../../store/store';
+import { userActions } from '../../store/user.slice';
 import LinkedMenuItem from '../LinkedMenuItem';
 import MenuIconBuilder from '../shared/MenuIconBuilder';
 import SimpleMenuItem from '../SimpleMenuItem';
 import styles from './index.module.css';
 
 function Menu() {
-	const {currentUser, isAuthenticated, logoutCurrentUser} = useContext(UserContext) as IUserContext;
+	const currentUser = useSelector((s: RootState) => s.users.currentUser);
+	const favors = useSelector((f: RootState) => f.favors.favorites);
+	const dispatch = useDispatch();
+
+	const logoutCurrentUser = () => {
+		if (currentUser) {
+			dispatch(userActions.logout(currentUser));
+		}
+	};
+
+
 	const authMenu = (
 		<>
 			<LinkedMenuItem
@@ -15,12 +26,12 @@ function Menu() {
 			/>
 			<LinkedMenuItem
 				text="Мои фильмы"
-				icon={MenuIconBuilder.buildCounterIcon(3)}
+				icon={MenuIconBuilder.buildCounterIcon(favors.length)}
 				linkUrl="/favorites"
 				handleClicks={(event) => console.log(event)}
 			/>
 			<SimpleMenuItem 
-				text={currentUser as string} /** if currentUser not null */
+				text={currentUser?.name as string} /** if currentUser not null */
 				icon={MenuIconBuilder.buildAvatarIcon()}
 				canChoose={false}
 			/>
@@ -38,7 +49,7 @@ function Menu() {
 			/>
 			<LinkedMenuItem
 				text="Мои фильмы"
-				icon={MenuIconBuilder.buildCounterIcon(3)}
+				icon={MenuIconBuilder.buildCounterIcon(0)}
 				linkUrl="/favorites"
 				handleClicks={(event) => console.log(event)}
 			/>
@@ -50,7 +61,7 @@ function Menu() {
 	);
 	return (
 		<ul className={styles.menu}>
-			{isAuthenticated ? authMenu: defaultMenuItems}
+			{currentUser ? authMenu: defaultMenuItems}
 		</ul>
 	);
 }