+++ search-films/src/context/user.context/index.tsx
@@ -1,5 +1,0 @@
-import UserContext from './user.context';
-import UserProvider from './user.provider/user.provider';
-export type { IUserContext } from './user.context';
-export { UserContext, UserProvider };
-