+++ search-films/CHANGELOG.md
@@ -1,3 +1,28 @@
+#### 0.4.0 (2024-11-18)
+
+##### Chores
+
+- delete deprecated code (6f05b117)
+- delete context from app (49effc4a)
+- add TODO (88809265)
+- **store:**
+  - add draft to configure store (5926b3b3)
+  - add storage.ts to load and save data (a9860c92)
+
+##### New Features
+
+- add util function to get duration (15af2af4)
+- add counter of favour films (e59887a1)
+- **user.slice:** add user.slice (need to test) (ea751c63)
+- **MovieCard:**
+  - add/delete WishList (2f8cd391)
+  * impl of add/remove from favour (5f9266a3)
+  * add util component MovieCard, minimal refactoring (a652f0d6)
+
+##### Refactors
+
+- minimal, delete useless code (7237fb03)
+
 #### 0.3.1 (2024-11-13)
 
 ##### New Features